Move CSVMapper to ocds_csv_mapper.py. Move base_json definition to view cove_ocds.views.explore_ocds()
